Replace the Error class with a fatal_msgbox function.
This class was ridiculous, aborting in its constructor. We even hit the heap, not even using an unnamed object or putting it on the stack! What we really wanted here is a "function". Also try to give it a "no return" attribute, to help static analysis (without it, there are false positive for "uninitialized variable").